Address PR review feedback: honest scale storage, weighted bpw, nits
Codex findings (both confirmed and fixed):
- Scales were stored fp32 while packed_state_bytes charged 2 bytes each
  and bit_budget claimed 16 bits. Scales (main + residual) are now
  rounded to fp16 at pack time when the budget says 16 -- quantisation
  runs against the rounded values so indices and dequant agree -- with a
  smallest-normal-fp16 floor so all-zero groups cannot underflow to a
  zero scale and NaN. Non-16-bit budgets keep fp32 and are charged their
  true element size.
- bits_per_weight_mean is now total stored bits over total quantised
  weights instead of an unweighted per-layer average, so small layers
  with high per-in_features overhead (TurboQuant per-row scales) no
  longer skew the model-wide rate used for equal-bit comparisons.

Copilot nits:
- Build LearnedRotation's tril indices directly on the target device via
  the device= keyword instead of a positional .to(device).
- Report rotation-training "steps" as an int.

77 tests pass (3 new: stored-scale precision, fp16 zero-group floor,
size-weighted footprint).

+++ rotquant/train_rotation.py
@@ -84,4 +84,4 @@ def train_layer_rotation(rotation: LearnedRotation, weight: torch.Tensor,
     rotation.eval()
     rotation.matrix()  # warm the eval-mode cache with the trained theta
     return {"initial_mse": initial, "final_mse": current_mse(),
-            "steps": float(cfg.steps)}
+            "steps": int(cfg.steps)}
